[Refactor] Weapon System Refactoring
