+++ survivor.io/Assets/Prefab/Weapon/Dagger.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 4338058672416997329}
   - component: {fileID: 1865286917341333024}
   - component: {fileID: 3732232483502519604}
+  - component: {fileID: 3484241063459111581}
   m_Layer: 0
   m_Name: Dagger
   m_TagString: Untagged
@@ -123,3 +124,15 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 1.3
+--- !u!114 &3484241063459111581
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1214319136903887294}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b80c66f44142e4b4db7f9fb5eb7a5a10, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
